+++ src/shared/config/index.tsx
@@ -1,0 +1,1 @@
+export { buttonVariants } from './variants';
