+++ src/shared/config/variants.ts
@@ -1,7 +1,6 @@
-import { cva, VariantProps } from 'class-variance-authority';
-import { forwardRef } from 'react';
+import { cva } from 'class-variance-authority';
 
-const buttonVariants = cva(
+export const buttonVariants = cva(
   'inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background',
   {
     variants: {
@@ -26,17 +25,3 @@ const buttonVariants = cva(
     },
   },
 );
-
-interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  className?: string;
-}
-
-export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant, size, ...props }, ref) => {
-    return <button className={buttonVariants({ variant, size, className })} ref={ref} {...props} />;
-  },
-);
-
-Button.displayName = 'Button';
