--- src/shared/ui/Card/index.ts
@@ -1,4 +1,0 @@
-export * from './Card';
-export * from './CardHeader';
-export * from './CardTitle';
-export * from './CardContent';
