--- src/shared/ui/Dialog/index.ts
@@ -1,4 +1,0 @@
-// export * from './Dialog';
-export * from './DialogHeader';
-export * from './DialogTitle';
-export * from './DialogContent';
