--- src/shared/ui/Select/SelectTrigger.tsx
@@ -1,50 +1,0 @@
-import { cva, VariantProps } from 'class-variance-authority';
-import { ChevronDown } from 'lucide-react';
-import { forwardRef } from 'react';
-import * as SelectPrimitive from '@radix-ui/react-select';
-
-const buttonVariants = cva(
-  'inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background',
-  {
-    variants: {
-      variant: {
-        default: 'bg-blue-500 text-white hover:bg-blue-600',
-        destructive: 'bg-red-500 text-white hover:bg-red-600',
-        outline: 'border border-gray-300 bg-transparent text-gray-700 hover:bg-gray-100',
-        secondary: 'bg-gray-200 text-gray-800 hover:bg-gray-300',
-        ghost: 'bg-transparent text-gray-700 hover:bg-gray-100',
-        link: 'underline-offset-4 hover:underline text-blue-500',
-      },
-      size: {
-        default: 'h-10 py-2 px-4',
-        sm: 'h-8 px-3 rounded-md text-xs',
-        lg: 'h-11 px-8 rounded-md',
-        icon: 'h-9 w-9',
-      },
-    },
-    defaultVariants: {
-      variant: 'default',
-      size: 'default',
-    },
-  },
-);
-
-interface SelectTriggerProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  className?: string;
-}
-
-export const SelectTrigger = forwardRef<HTMLButtonElement, SelectTriggerProps>(
-  ({ className, children, ...props }, ref) => (
-    <SelectPrimitive.Trigger
-      ref={ref}
-      className={`flex h-10 items-center justify-between rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-      {...props}
-    >
-      {children}
-      <ChevronDown className='h-4 w-4 opacity-50' />
-    </SelectPrimitive.Trigger>
-  ),
-);
-SelectTrigger.displayName = SelectPrimitive.Trigger.displayName;
