--- src/shared/ui/Select/index.ts
@@ -1,4 +1,0 @@
-export * from './Select';
-export * from './SelectContent';
-export * from './SelectItem';
-export * from './SelectTrigger';
