--- src/shared/ui/Table/TableCell.tsx
@@ -1,16 +1,0 @@
-import { forwardRef } from 'react';
-
-interface TableCellProps extends React.HTMLAttributes<HTMLTableCellElement> {
-  className?: string;
-}
-
-export const TableCell = forwardRef<HTMLTableCellElement, TableCellProps>(
-  ({ className, ...props }, ref) => (
-    <td
-      ref={ref}
-      className={`p-2 align-middle [&:has([role=checkbox])]:pr-0 ${className}`}
-      {...props}
-    />
-  ),
-);
-TableCell.displayName = 'TableCell';
