--- src/shared/ui/Table/TableHeader.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { forwardRef } from 'react';
-
-interface TableHeaderProps extends React.HTMLAttributes<HTMLTableSectionElement> {
-  className?: string;
-}
-
-export const TableHeader = forwardRef<HTMLTableSectionElement, TableHeaderProps>(
-  ({ className, ...props }, ref) => (
-    <thead ref={ref} className={`[&_tr]:border-b ${className}`} {...props} />
-  ),
-);
-TableHeader.displayName = 'TableHeader';
