--- src/shared/ui/Table/index.ts
@@ -1,6 +1,0 @@
-export * from './Table';
-export * from './TableBody';
-export * from './TableCell';
-export * from './TableHead';
-export * from './TableHeader';
-export * from './TableRow';
